+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # callr development version
 
+* callr now handles symbolic arguments properly, and does not evaluate them.
+  E.g. `callr::r(function(x) x, list(quote(foobar)))` works now (#175).
+
 # callr 3.5.1
 
 * `callr::r_session` now handles large messages from the subprocess